Clean up test driver code some more
