+++ lib/nodejs/test/client.js
@@ -24,8 +24,8 @@ var assert = require('assert');
 var thrift = require('thrift');
 var helpers = require('./helpers');
 var ThriftTest = require('./gen-nodejs/ThriftTest');
-var ThriftTestDriver = require('./thrift_test_driver').ThriftTestDriver;
-var ThriftTestDriverPromise = require('./thrift_test_driver_promise').ThriftTestDriver;
+var ThriftTestDriver = require('./test_driver').ThriftTestDriver;
+var ThriftTestDriverPromise = require('./test_driver').ThriftTestDriverPromise;
 var SecondService = require('./gen-nodejs/SecondService');
 var ttypes = require('./gen-nodejs/ThriftTest_types');
